fix(codex): refresh codegraph runtime gate fixes
